fast_lock: perform wait for cleaning markdirs in parallel
For each markdir present a separate 2s wait was being
performed.  Now find all the present markdirs, wait 2s
and then clean them all up.

This means that some slow creators which previously may
have not been interrupted could now get interrupted.  Those
slow creators previously immune to this were not intentionally
immune to this.  All slow creators have always faced the risk
of being interrupted, it just may happen earlier now.  With this
fix, slow creators are given a 2s window to perform a complete
mkdir -p.  This operation can be slow on a cross-site NFS server.
Testing shows that the entire fast_lock between SD and Boulder
can takes around 1s.  If this ends up causing a problem, the
wait period can be increased for all markers simultaneously
instead of the old policy that allows some to be shorter and
some to be longer.

+++ fast_lock.sh
@@ -70,13 +70,16 @@ delete_marker() { # lock uid
 marker_ids() { ( shopt -s nullglob ; basenames "$1/markers"/* ) }
 
 clean_markers() { # lock
-    local lock=$1  mid ; shift
-    for mid in $(marker_ids "$lock") ; do
-        sleep 2
-        # deleting marker dirs without a marker can cause really
-        # slow creates to fail (oh well, they will have to try
-        # again, locking is a privellege not a right).  This helps
-        # remove stale dirs without a marker.
+    local lock=$1 mids
+    mids=($(marker_ids "$lock"))
+    [ -z "$mids" ] && return
+    sleep 2
+    # deleting marker dirs without a marker can cause really
+    # slow creates to fail (oh well, they will have to try
+    # again, locking is a privellege not a right).  This helps
+    # remove stale dirs without a marker.
+    local mid
+    for mid in "${mids[@]}" ; do
         delete_markdirs "$lock" "$mid"
     done
     q rmdir "$lock"
